+++ tests/test_integrations.py
@@ -3,6 +3,7 @@ from django.urls import reverse
 
 from cms.api import add_plugin, create_page, create_title
 from cms.test_utils.testcases import CMSTestCase
+from cms.toolbar.utils import get_object_preview_url
 
 from djangocms_alias.models import Alias as AliasModel, AliasContent, Category
 from djangocms_alias.utils import is_versioning_enabled
@@ -112,3 +113,6 @@ class NestedAppIntegrationTestCase(CMSTestCase):
             response = self.client.get(references_endpoint)
 
         self.assertContains(response, poll_content)
+        self.assertContains(response, "pagecontent")
+        self.assertContains(response, get_object_preview_url(page_content))
+        self.assertContains(response, page_content.versions.first().state)